pyproject.toml
[project]
name = "oc-broker"
version = "0.1.0"
description = "OpenClaw Remote Broker MCP - Single-user remote execution gateway over Model Context Protocol"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "cryptography>=41.0.0",
    "structlog>=23.2.0",
    "psutil>=5.9.6",
    "aiofiles>=23.2.0",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "ruff>=0.1.6",
]

[project.scripts]
oc-broker = "main:main"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
